requirements.txt
alembic==1.16.5
black==25.9.0
blinker==1.9.0
cfgv==3.4.0
click==8.3.0
contourpy==1.3.3
cycler==0.12.1
distlib==0.4.0
dnspython==2.8.0
email-validator==2.3.0
filelock==3.19.1
flake8==7.3.0
Flask==3.1.2
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fonttools==4.60.0
greenlet==3.2.4
gunicorn==23.0.0
identify==2.6.14
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.6
mccabe==0.7.0
mypy_extensions==1.1.0
nodeenv==1.9.1
numpy==2.3.3
packaging==25.0
pandas==2.3.2
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.6.0
pre_commit==4.3.0
psycopg2-binary==2.9.10
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
pyparsing==3.2.5
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytokens==0.1.10
pytz==2025.2
PyYAML==6.0.3
six==1.17.0
SQLAlchemy==2.0.43
typing_extensions==4.15.0
tzdata==2025.2
virtualenv==20.34.0
Werkzeug==3.1.3
WTForms==3.2.1
